fold showQuickPick into showWorkflowPick
